Split XHTM from preprocess in integration tests
